PR feedback incorporated in the code
