+++ helpers/types.ts
@@ -10,7 +10,6 @@ export type ModelProvider =
   | "t-systems";
 export type ModelConfig = {
   provider: ModelProvider;
-  apiBase?: string;
   apiKey?: string;
   model: string;
   embeddingModel: string;
